+++ src/Federation/SchemaPrinter.php
@@ -9,7 +9,6 @@ use GraphQL\Type\Definition\InterfaceType;
 use GraphQL\Type\Definition\ObjectType;
 use GraphQL\Type\Definition\Type;
 use GraphQL\Utils\SchemaPrinter as GraphQLSchemaPrinter;
-use GraphQL\Utils\Utils;
 
 class SchemaPrinter extends GraphQLSchemaPrinter
 {
@@ -105,9 +104,12 @@ GRAPHQL;
         return ' '
             . implode(
                 ' ',
-                Utils::map($directives, static function (DirectiveNode $directive): string {
-                    return Printer::doPrint($directive);
-                })
+                array_map(
+                    static function (DirectiveNode $directive): string {
+                        return Printer::doPrint($directive);
+                    },
+                    $directives
+                )
             );
     }
 }