+++ src/Schema/Directives/ThrottleDirective.php
@@ -135,7 +135,7 @@ GRAPHQL;
             // @phpstan-ignore-next-line won't be executed on Laravel < 8
             $limiter = $this->limiter->limiter($name);
             // @phpstan-ignore-next-line $limiter may be null although it's not specified in limiter() PHPDoc
-            if (is_null($limiter)) {
+            if (null === $limiter) {
                 throw new DefinitionException("Named limiter {$name} is not found.");
             }
         }